Changes to entity energy management:
- change energy data type to float
- initialize entity energy & base energy to default value
- use base energy attribute when calculating energy recovery

+++ include/Entity.hpp
@@ -41,12 +41,8 @@ protected:
 	/** Rate of momentum. */
 	float momentum = 0;
 
-	/**
-	 * Entity's current energy level.
-	 *
-	 * FIXME: change to float?
-	 */
-	int32_t energy;
+	/** Entity's current energy level. */
+	float energy;
 
 public:
 	/**
